switch vm stack to little endian representation
H256 is [u8; 32]
U256 is [u64; 4]

According to EVM spec, when H256 is interpreted as a number, the byte
order is big endian. WASM, in contrast, uses little-endian
representation (the same used by CPUs internally).

Before this PR, evm stored operand stack as `Vec<H256>`, so any
arithmetic op entailed indianness conversion. As WASM lacks equivalent
to bswap instruction, this is a lot of work.

After this PR, stack is using U256 (so, little-endian in memory). This
is also reperesentation used by all of geth, evmone and odin.

For the uniswap benchmark, the before-after WASM (not total) gas cost is

110436615921984
 95290678433280