Use usize rather than u256 for memory offsets
This change improves performance (as measured by the number of WASM
instructions executed) by about 5%.

The story here is that memory-related operations, notably mload and
mstore, are pretty high in profiles. Part of the reason for that is that
offsets are represented as U256, and carrying out bounds checking in
wide integers is naturally slow!

Luckily, we actually know that all actual addresses will fit in usize --
bigger addresses *must* run out of gas. So, in this PR we do two things:

* when calculating memory-related gas usage, convert u256 to usize as
  early as possible, flagging overflows as out of gas errors.
* when executing memory-related instructions, *assume* that operands fit
  usize and just cast them (we use checked cast, so we'll panic if
  that's not the case. Replacing that with unchecked cast leads to just
  a tiny further perf improvement).

I am not entirely thrilled by this non-local invariant. It'd be much
better if gas counting and execution were defined by the same function,
so as to make the invariant local. Still, I think the current impl is
fine:

Even in the original impl, we had a similar implicit invariant
(as_usize_or_fail was casting u256 to usize). The official geth impl
seems to do roughly the same thing. As far as I understand, they even
use unchecked casts!